Dev Kit: Another long to int32_t.
git-svn-id: svn://svn.code.sf.net/p/opensludge/code@261 27a91dd4-ec73-4ff0-838b-64dfa05f697e

+++ source/ProjectManager/settings.cpp
@@ -2,6 +2,7 @@
 #include <dirent.h>
 #include <string.h>
 #include <stdio.h>
+#include <stdint.h>
 #include <unistd.h>
 #include <sys/types.h>
 #include <sys/stat.h>
@@ -305,8 +306,8 @@ FILE * openFinalFile (char * addMe, char * mode) {
 }
 
 typedef struct _FILETIME {
-	unsigned long dwLowDateTime;
-	unsigned long dwHighDateTime;
+	uint32_t dwLowDateTime;
+	uint32_t dwHighDateTime;
 } FILETIME;
